requirements.txt
streamlit==1.29.0
alpaca-py==0.13.3
ccxt==4.1.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
python-dotenv==1.0.0
